+++ src/main/java/com/coupang/marketplace/user/controller/UserController.java
@@ -1,17 +1,14 @@
 package com.coupang.marketplace.user.controller;
 
-import static com.coupang.marketplace.auth.LoginType.*;
-
-import com.coupang.marketplace.auth.LoginAuth;
 import com.coupang.marketplace.global.common.StatusEnum;
 import com.coupang.marketplace.global.common.SuccessResponse;
+import com.coupang.marketplace.user.controller.dto.SignInRequestDto;
+import com.coupang.marketplace.user.controller.dto.SignUpRequestDto;
 import com.coupang.marketplace.user.service.LoginService;
 import com.coupang.marketplace.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -47,16 +44,5 @@ public class UserController {
     @GetMapping("/users/logout")
     public void logoutUser(){
         loginService.logout();
-    }
-
-    @LoginAuth(type = USER)
-    @PutMapping("/mypage/user-information/{id}")
-    public SuccessResponse updateUser(@PathVariable Long id, @Valid @RequestBody UpdateRequestDto requestDto){
-        userService.updateUser(id, requestDto);
-        SuccessResponse res = SuccessResponse.builder()
-            .status(StatusEnum.OK)
-            .message("회원 정보 수정 성공")
-            .build();
-        return res;
     }
 }
