+++ src/main/java/com/coupang/marketplace/user/controller/dto/SignInRequestDto.java
@@ -1,4 +1,4 @@
-package com.coupang.marketplace.user.controller;
+package com.coupang.marketplace.user.controller.dto;
 
 import lombok.Builder;
 import lombok.Getter;
