+++ src/main/java/com/coupang/marketplace/user/service/LoginService.java
@@ -1,6 +1,6 @@
 package com.coupang.marketplace.user.service;
 
-import com.coupang.marketplace.user.controller.SignInRequestDto;
+import com.coupang.marketplace.user.controller.dto.SignInRequestDto;
 
 public interface LoginService {
