+++ src/main/java/com/coupang/marketplace/user/service/SessionLoginService.java
@@ -1,6 +1,6 @@
 package com.coupang.marketplace.user.service;
 
-import com.coupang.marketplace.user.controller.SignInRequestDto;
+import com.coupang.marketplace.user.controller.dto.SignInRequestDto;
 import com.coupang.marketplace.user.domain.User;
 import com.coupang.marketplace.user.repository.UserRepository;
 import com.coupang.marketplace.global.constant.SessionKey;
