+++ src/test/java/com/coupang/marketplace/user/service/SessionLoginServiceTest.java
@@ -16,7 +16,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.coupang.marketplace.user.controller.SignInRequestDto;
+import com.coupang.marketplace.user.controller.dto.SignInRequestDto;
 import com.coupang.marketplace.user.domain.User;
 import com.coupang.marketplace.fixture.UserFixture.*;
 import com.coupang.marketplace.user.repository.UserRepository;
